slack_bot/requirements.txt
openai==0.27.8
slack-bolt==1.18.0
slack-sdk==3.21.3
langchain==0.0.194
sentence-transformers==2.2.2
pinecone-client[grpc]==2.2.2
wandb==0.15.9
